agentstatus/requirements.txt
PyYAML
Requests
urllib3
cefevent